Fix shadowing of `lower_bound` and `upper_bound`, and pass the comparator.
I don't know what is the Standardese style here: normally an algorithm
in the `std` namespace would not be qualified, but here we have to do
something to distinguish the `std::lower_bound` algorithm from the
`flat_set::lower_bound` member function which in reality would hide the
name of the algorithm.

Separately, of course we need to pass `compare` here.

Notice that we do NOT pass `compare` to `std::lexicographical_compare`
in the definition of the comparison operators. This is a bit bizarre,
but it follows the precedent set by `std::set` and `std::map`.