Wording for `extract()` now specifies something not-broken for PMR containers.
The old wording was in terms of "default-construct a KeyContainer `temp`,
then swap it with `c`." The swap will have undefined behavior whenever
`temp`'s default-constructed allocator is not equal to `c`'s allocator.
This is generally expected to be the case when KeyContainer is anything
with `std::pmr` in the name (e.g. `std::pmr::vector`).

    std::pmr::monotonic_buffer_resource mr;
    std::pmr::polymorphic_allocator<int> a(&mr);
    std::flat_set<int, std::less<>, std::pmr::vector<int>> fs({1, 2}, a);
    std::pmr::vector<int> v = std::move(fs).extract();
    assert(v.get_allocator() == a);